Refactor UpdateUser to use Redis Cache for reducing database load
This commit refactors the UpdateUser function in order to reduce the database load by using Redis Cache for caching the updated user.

The changes include:

Using id parameter as an integer instead of a string in the UpdateUser function.
Removing unnecessary error return from updateUserInDBAndCache function.
Removing unnecessary Rollback() in updateUserInDBAndCache function.
Removing redundant password hashing in updateUserInDBAndCache function.
These changes will help to improve the performance and scalability of the application, by reducing the load on the database and utilizing Redis cache for faster access to frequently requested data.